Add files via upload
Updated for coding standards and feedback to add exit sequence in Saloon.java. Deleted "tapping game" text out of the Saloon Print Statement in BoomTown.java now that the final version of saloon is working. Also added more drinks into the drinks list to help randomize the orders that the user must fill.